# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Events Management LLC API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://events-management.test'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
...
  '/api/bans/{user}':
    get:
      summary: 'Users Banned'
      operationId: usersBanned
      description: "List all the users banned by the authenticated user or a specified user.\n* Only administrators can view other users' banned lists."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 7553
                      name: 'Morgan Hirthe'
                      email: dare.emelie@example.com
                      country: Peru
                      profession: 'Compliance Officers'
                      phone: 562.471.7515
                      organization: Mitchell-VonRueden
                    -
                      id: 7554
                      name: 'Dr. Grayson Glover'
                      email: schultz.audrey@example.org
                      country: 'French Polynesia'
                      profession: 'Extraction Worker'
                      phone: +1.518.885.7351
                      organization: 'Fahey, Cartwright and Balistreri'
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 7553
                        name: 'Morgan Hirthe'
                        email: dare.emelie@example.com
                        country: Peru
                        profession: 'Compliance Officers'
                        phone: 562.471.7515
                        organization: Mitchell-VonRueden
                      -
                        id: 7554
                        name: 'Dr. Grayson Glover'
                        email: schultz.audrey@example.org
                        country: 'French Polynesia'
                        profession: 'Extraction Worker'
                        phone: +1.518.885.7351
                        organization: 'Fahey, Cartwright and Balistreri'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 7553
                        name:
                          type: string
                          example: 'Morgan Hirthe'
                        email:
                          type: string
                          example: dare.emelie@example.com
                        country:
                          type: string
                          example: Peru
                        profession:
                          type: string
                          example: 'Compliance Officers'
                        phone:
                          type: string
                          example: 562.471.7515
                        organization:
                          type: string
                          example: Mitchell-VonRueden
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: "You are not authorized to view this user's banned list."
                properties:
                  message:
                    type: string
                    example: "You are not authorized to view this user's banned list."
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Event not found.'
                properties:
                  message:
                    type: string
                    example: 'Event not found.'
      tags:
        - Bans
    parameters:
      -
        in: path
        name: user
        description: "Optional parameter. The ID of the user whose banned list to retrieve. If not specified, retrieves the authenticated user's banned list."
        required: true
        schema:
          type: integer
        examples:
          omitted:
            summary: 'When the value is omitted'
            value: ''
          present:
            summary: 'When the value is present'
            value: '1'
...